+++ R/compute_aic.R
@@ -3,9 +3,10 @@ globalVariables(c(glm.nb, AIC, glm, gaussian, glm.control))
 #' model and the Voom model
 #'
 #' This function calculates the AIC based on lognormal distribution,
-#' negative-binomial distribution as well as the voom transformation.
+#' negative binomial distribution as well as the voom transformation.
 #' It then compares the AICs of the three models across different genes and
-#' yields AIC-based metric values.
+#' yields AIC-based metric values. Must have non-negative data and non-discrete
+#' data will only analyze the lognormal and voom distribution.
 #'
 #' @param se SummarizedExperiment object
 #' @param assay_of_interest The assay name from se that you are interested in
@@ -30,16 +31,16 @@ globalVariables(c(glm.nb, AIC, glm, gaussian, glm.control))
 #'     model for individual genes.}
 #'     \item{voom_result}{A vector contains the AIC based on voom transformation
 #'     for individual genes.}
-#'     \item{total_AIC}{The sum of AICs across all genes for the three models
-#'     in comparison.}
-#'     \item{min_AIC}{The number of minimum AIC across the three models in
+#'     \item{total_AIC}{The sum of AICs across all genes for the models in
+#'     comparison.}
+#'     \item{min_AIC}{The number of minimum AIC across the models in
 #'     comparison for individual genes.}
-#'     \item{AIC_metric}{The ratios between total_AIC and min_AIC for the three
+#'     \item{AIC_score}{The ratios between total_AIC and min_AIC for the
 #'     models in comparison. The optimal model is the one has minimum value.}
 #'   }
 #'
 #' @return A list with a dataframe of the total_AIC values, a dataframe with
-#' the min_AIC values, and AIC_metric vector with the following three elements:
+#' the min_AIC values, and AIC_score vector with the following three elements:
 #' \describe{
 #'     \item{nb}{The metric value calculated based on the negative binomial
 #'       model.}
@@ -68,19 +69,18 @@ compute_aic <- function(se, assay_of_interest, batchind,
     design <- stats::model.matrix(stats::as.formula(paste(" ~",
                             paste(colnames(analysis_design), collapse = "+"))),
                             data = analysis_design)
-    if (!all(dat == floor(dat)) || any(dat < 0)) {
-        stop("Counts must be non-negative integers only.")
+    nb_test <- TRUE
+    num_models <- 2
+
+    if (any(dat < 0)) {
+        stop("Counts must be non-negative values only.")
+    } else if (!all(dat == floor(dat))){
+        nb_test <- FALSE
     }
+
     dat <- dat[rowSums(dat) != 0, ]
     dat <- dat[rowSums(dat == 0) <= zero_filt_percent / 100 * ncol(dat), ]
-    nb_result <- apply(dat, 1, function(x) {
-        tryCatch({
-        nb_model <- glm.nb(x ~ design, control = glm.control(maxit = maxit))
-        nb_AIC <- AIC(nb_model)
-        return(nb_AIC)
-    }, error = function(e) {
-        return(NA)
-    })})
+
     lognormal_result <- apply(dat, 1, function(x) {
         tryCatch({
         lognormal_model <- glm(log(x + 1e-100) ~ design, family = gaussian)
@@ -89,6 +89,7 @@ compute_aic <- function(se, assay_of_interest, batchind,
     }, error = function(e) {
         return(NA)
     })})
+
     voom_dat <- voom(dat, design = design)
     voom_dat <- voom_dat$E
     voom_result <- apply(voom_dat, 1, function(x) {
@@ -99,15 +100,31 @@ compute_aic <- function(se, assay_of_interest, batchind,
     }, error = function(e) {
         return(NA)
     })})
-    aic_matrix <- cbind(nb_result, lognormal_result, voom_result)
-    colnames(aic_matrix) <- c("nb_AIC", "lognormal_AIC", "voom_AIC")
-    total_AIC <- colSums(aic_matrix, na.rm = TRUE)
+
+    if(nb_test){
+        nb_result <- apply(dat, 1, function(x) {
+            tryCatch({
+                nb_model <- glm.nb(x ~ design, control = glm.control(maxit = maxit))
+                nb_AIC <- AIC(nb_model)
+                return(nb_AIC)
+            }, error = function(e) {
+                return(NA)
+            })})
+        aic_matrix <- cbind(nb_result, lognormal_result, voom_result)
+        colnames(aic_matrix) <- c("nb_AIC", "lognormal_AIC", "voom_AIC")
+        num_models <- 3
+    } else {
+        aic_matrix <- cbind(lognormal_result, voom_result)
+        colnames(aic_matrix) <- c("lognormal_AIC", "voom_AIC")
+    }
+
+    total_AIC <- as.data.frame(t(colSums(aic_matrix, na.rm = TRUE)), colnames = "")
     min_model <- apply(aic_matrix, 1, function(x) {
         if (all(is.na(x))) return(NA)
         which.min(x)
     })
-    min_AIC <- table(factor(min_model, levels = seq_along(seq(1, 3))))
-    names(min_AIC) <- c("nb", "lognormal", "voom")
+    min_AIC <- t(as.data.frame(table(factor(min_model, levels = seq_along(seq(1, num_models)))))[2])
+    colnames(min_AIC) <- colnames(aic_matrix)
     return(list(total_AIC = total_AIC, min_AIC = min_AIC,
-        AIC_metric = total_AIC / min_AIC))
+        AIC_score = as.data.frame(total_AIC/min_AIC, row.names = "")))
 }